Refactor form submit and update into workflow_inputs
